+++ disney-app/package.json
@@ -15,6 +15,7 @@
     "@types/react-dom": "^18.0.11",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.8",
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "sass": "^1.59.3",
@@ -68,10 +69,10 @@
     "webpack": "^5.76.2",
     "webpack-cli": "^5.0.1"
   },
-  "jest": { 
-    "collectCoverageFrom": [ 
-        "src/**/*.tsx", 
-        "!**/node_modules/**"
-    ] 
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.tsx",
+      "!**/node_modules/**"
+    ]
   }
 }
