+++ disney-app/src/components/notification.tsx
@@ -14,10 +14,9 @@ const StyledMessage = styled.div`
 interface IMessageProps {
   message: string;
 }
-class Message extends React.Component<IMessageProps> {
-  render() {
-    return <StyledMessage data-testid="message">{this.props.message}</StyledMessage>;
-  }
-}
+
+const Message = (props: IMessageProps) => {
+  return <StyledMessage data-testid="message">{props.message}</StyledMessage>;
+};
 
 export default Message;
