+++ disney-app/src/components/oneNewCard.tsx
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import styled from '@emotion/styled';
 import { INewHero } from 'data/HPResponse.models';
 
@@ -22,25 +22,22 @@ const StyledCard = styled.div`
 interface IOneNewCardProps {
   card: INewHero;
 }
-class OneNewCard extends Component<IOneNewCardProps> {
-  render() {
-    const image = this.props.card.image;
-    if (image) {
-      const imageUrl = URL.createObjectURL(image);
-      return (
-        <StyledCard data-testid="new-card">
-          <h3>
-            <b>{this.props.card.name}</b>
-          </h3>
-          <p>Date of Birth: {this.props.card.dateOfBirth}</p>
-          <p>Wizard: {String(this.props.card.wizard)}</p>
-          <p>House: {this.props.card.house}</p>
-          <p>Gender: {this.props.card.gender}</p>
-          <img src={imageUrl} alt="newHero" />
-        </StyledCard>
-      );
-    }
-  }
-}
+
+const OneNewCard = (props: IOneNewCardProps) => {
+  const image = props.card.image;
+  const imageUrl = URL.createObjectURL(image as File);
+  return (
+    <StyledCard data-testid="new-card">
+      <h3>
+        <b>{props.card.name}</b>
+      </h3>
+      <p>Date of Birth: {props.card.dateOfBirth}</p>
+      <p>Wizard: {String(props.card.wizard)}</p>
+      <p>House: {props.card.house}</p>
+      <p>Gender: {props.card.gender}</p>
+      <img src={imageUrl} alt="newHero" />
+    </StyledCard>
+  );
+};
 
 export default OneNewCard;
