+++ disney-app/src/pages/AboutPage.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 
-class AboutPage extends React.Component {
-  render() {
-    return (
-      <h1 data-testid="about-page">
-        Тут будет информация по типу about. Что писать, я пока не придумала, но потом сделаю
-        что-нибудь
-      </h1>
-    );
-  }
-}
+const AboutPage = () => {
+  return (
+    <h1 data-testid="about-page">
+      Тут будет информация по типу about. Что писать, я пока не придумала, но потом сделаю
+      что-нибудь
+    </h1>
+  );
+};
 
 export default AboutPage;
