+++ disney-app/src/pages/HomePage.tsx
@@ -1,15 +1,12 @@
 import AllCards from 'containers/AllCards';
 import response from 'data/HPResponse';
-import React, { Component } from 'react';
+import React from 'react';
 
-class HomePage extends Component {
-  render() {
-    return (
-      <div data-testid="home-page">
-        <AllCards cards={response} />;
-      </div>
-    );
-  }
-}
-
+const HomePage = () => {
+  return (
+    <div data-testid="home-page">
+      <AllCards cards={response} />;
+    </div>
+  );
+};
 export default HomePage;
